fixed some comments
