+++ src/ra/CMakeLists.txt
@@ -8,9 +8,9 @@ CREATE_RA_TEST(all_test)
 CREATE_RA_TEST(count_test)
 CREATE_RA_TEST(cross_test)
 CREATE_RA_TEST(filter_test)
+CREATE_RA_TEST(groupavg_test)
+CREATE_RA_TEST(groupcount_test)
+CREATE_RA_TEST(groupsum_test)
 CREATE_RA_TEST(iterable_test)
 CREATE_RA_TEST(map_test)
 CREATE_RA_TEST(project_test)
-CREATE_RA_TEST(groupsum_test)
-CREATE_RA_TEST(groupcount_test)
-CREATE_RA_TEST(groupavg_test)
